receiver/requirements.txt
azure-identity
azure-servicebus
azure-data-tables
requests
